% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themers.R
\name{active_color_theme}
\alias{active_color_theme}
\title{Get The Name Of The Active Color Theme}
\usage{
active_color_theme()
}
\value{
(character) color theme name
}
\description{
Get The Name Of The Active Color Theme
}
\examples{
active_color_theme()
}
