% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\name{add_asofdate}
\alias{add_asofdate}
\title{Add Export As of Date}
\usage{
add_asofdate(wb, x, sheet = current_sheet(), row = 3, doc_width = "G")
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{x}{(character or Date time) The Date (time) when the Export data
was created}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{row}{in which row the As Of Date should be added. Default = 3}

\item{doc_width}{(single character or numeric) Up to which column should the
texts spread? Default = "G"}
}
\value{
A wbWorkbook, invisibly
}
\description{
Add Export As of Date
}
\examples{
wb <- openxlsx2::wb_workbook()
style_datadict(wb)
wb$add_worksheet()
add_asofdate(wb, "2022-11-20")
}
