% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_form_overview.R
\name{add_form_overview}
\alias{add_form_overview}
\title{Create And Add A Form Overview To The Data Dictionary}
\usage{
add_form_overview(
  wb,
  datadict_tables,
  title = NULL,
  subtitle = NULL,
  as_of_date = NULL,
  form_type_description = TRUE,
  doc_width = "G"
)
}
\arguments{
\item{wb}{a workbook}

\item{datadict_tables}{a list with tables for the data dictionary,
created by \code{\link[=create_datadict_tables]{create_datadict_tables()}}}

\item{title}{(character) Title of the document. Default = NULL}

\item{subtitle}{(character) subtitle of the document,
e.g. researcher - study name.
Default = NULL'}

\item{as_of_date}{(date) when the data set was created. Default = NULL}

\item{form_type_description}{(logical), Should short explanation of the
different form types added to the form overview? Default = TRUE}

\item{doc_width}{(single character or numeric) Up to which column should the
titles and texts spread? Default = "G"}
}
\value{
invisible, the workbook with a form overview sheet added to it.
}
\description{
Create And Add A Form Overview To The Data Dictionary
}
