% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_form_sheets.R
\name{add_form_sheets}
\alias{add_form_sheets}
\title{Ceate And Add Form Sheets To The Data Dictionary}
\usage{
add_form_sheets(wb, datadict_tables, var_select = FALSE)
}
\arguments{
\item{wb}{a workbook}

\item{datadict_tables}{generated with \code{\link[=create_datadict_tables]{create_datadict_tables()}}}

\item{var_select}{(logical) Is this a variable selection document?}
}
\value{
workbook with form item tables added (invisibly)
}
\description{
Create a workbook sheet for every table in the datadict_tables$form_items
list.

The parameter "var_select" is ignored so far but in future release a 'select'
column with selection rules will be added to the form sheet tables, if
'var_select' = TRUE
}
\examples{
\dontrun{
wb <- wb_workbook()
add_form_overview(wb, datadict_tables = datadict_tables)
add_form_sheets(wb, datadict_tables = datadict_tables)
}
}
