% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\name{add_heading2}
\alias{add_heading2}
\title{Add Heading 2}
\usage{
add_heading2(wb, x, sheet = current_sheet(), row = 1, doc_width = "G")
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{x}{(character) The Heading text}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{row}{in which row the heading should be added. Default = 1}

\item{doc_width}{(single character or numeric) Up to which column should the
texts spread? Default = "G"}
}
\value{
A wbWorkbook, invisibly
}
\description{
Add Heading 2
}
\examples{
wb <- openxlsx2::wb_workbook()
style_datadict(wb)
wb$add_worksheet()
add_heading2(wb, "My Subheading")
}
