% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\name{add_paragraph}
\alias{add_paragraph}
\title{Add A text paragraph}
\usage{
add_paragraph(
  wb,
  x,
  sheet = current_sheet(),
  row = 1,
  doc_width = "G",
  font_weight = 0.86
)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{x}{(character string) the text of the paragraph}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{row}{in which row the As Of Date should be added. Default = 1}

\item{doc_width}{(single character or numeric) Up to which column should the
texts spread? Default = "G"}

\item{font_weight}{(numeric) correction factor if the estimated number of
needed lines is off. Default: 0.86}
}
\value{
A wbWorkbook, invisibly
}
\description{
This function merges cells of a given document with (e.g. A:G) and sets the
cell style to "text_area". Within this cell style the text is wrapped.
However Excel fails to set the correct row height, when text is added into
merged cells. This function tries to estimate the right number of lines needed
to display the text and set the row height (with some extra space at the end)
accordingly.
}
\examples{
wb <- openxlsx2::wb_workbook()
style_datadict(wb)
wb$add_worksheet()

p <- paste(
  "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy",
  "eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam",
  "voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet",
  "clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.")

add_paragraph(wb, p)
}
