% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_scopes.R
\name{add_scope_selection}
\alias{add_scope_selection}
\title{Add Scope Selection Table to workbook}
\usage{
add_scope_selection(
  wb,
  scope,
  title = NULL,
  sheet = current_sheet(),
  row = 1,
  doc_width = "G"
)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{scope}{"Scope Table", including at least a scopeID column named "Scope".}

\item{title}{(character) A title for the Scope selection table. Default = NULL
If title = NULL, no title will be created.}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{row}{number where to start the current Scope Selection, default = 1}

\item{doc_width}{(single character or numeric) if a title is given, up to which
column should the section title spread? Default = "G"}
}
\value{
wbWorkbook, invisibly
}
\description{
Add Scope Selection Table to workbook
}
\examples{
\dontrun{
add_scope_selection(
  wb,
  scope = datapkg_scopes,
  title = "Data Packages",
  row = 5,
  doc_width = "I"
)
}
}
