% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\name{add_styled_line}
\alias{add_styled_line}
\title{Add Styled Text To The Workbook}
\usage{
add_styled_line(
  wb,
  x,
  style,
  sheet = current_sheet(),
  row = 1,
  start_col = "A",
  doc_width = "G"
)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{x}{The text to add to the workbook}

\item{style}{(character) name of the cell style defined to the workbook.
If your worbook is named \code{wb} enter \code{wb$styles_mgr$xf$name} to get a list
of style names registered to the workbook}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{row}{(integer) in which row the title should be added. Default = 1}

\item{start_col}{(single character or numeric) From which column should the
texts start? Default = "A"}

\item{doc_width}{(single character or numeric) Up to which column should the
texts spread? Default = "G"}
}
\value{
A wbWorkbook, invisibly
}
\description{
Add some Text to a given row, merge the cells of that row and apply a
predefined style.
}
\examples{
wb <- openxlsx2::wb_workbook()
style_datadict(wb)
wb$add_worksheet()
add_styled_line(wb, "some text", style = "title")
}
