% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\name{add_subtitle}
\alias{add_subtitle}
\title{Add Document Subtitle}
\usage{
add_subtitle(wb, x, sheet = current_sheet(), row = 2, doc_width = "G")
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{x}{(character) The document subtitle}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{row}{in which row the subtitle should be added. Default = 2}

\item{doc_width}{(single character or numeric) Up to which column should the
texts spread? Default = "G"}
}
\value{
A wbWorkbook, invisibly
}
\description{
Add Document Subtitle
}
\examples{
wb <- openxlsx2::wb_workbook()
style_datadict(wb)
wb$add_worksheet()
add_subtitle(wb, "My Subtitle")
}
