% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\name{add_title}
\alias{add_title}
\title{Add Document Title}
\usage{
add_title(wb, x, sheet = current_sheet(), row = 1, doc_width = "G")
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{x}{(character) The document title}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{row}{(integer) in which row the title should be added. Default = 1}

\item{doc_width}{(single character or numeric) Up to which column should the
texts spread? Default = "G"}
}
\value{
A wbWorkbook, invisibly
}
\description{
Add Document Title
}
\examples{
wb <- openxlsx2::wb_workbook()
style_datadict(wb)
wb$add_worksheet()
add_title(wb, "My Workbook Title")
}
