% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\name{clean_region_name}
\alias{clean_region_name}
\title{Clean Excel Region Names}
\usage{
clean_region_name(name)
}
\arguments{
\item{name}{(character or character vector) the name(s) of the named region}
}
\value{
a string (vector) of cleaned region names.
}
\description{
Names for named regions have the following restrictions:
\itemize{
\item The first character of a name must be a letter or an underscore character (_).
\item Remaining characters in the name can be letters, numbers, periods, and
underscore characters.
\item Spaces are not allowed.
\item Names cannot be the same as a cell reference, such as Z$100, BIN9, or R1C1.
}
}
\details{
Therefore this function
\itemize{
\item replaces all punctuation characters like "-" with "."
\item replaces all white spaces with "_"
\item inserts an underscore at the beginning if
\itemize{
\item a name starts with anything else than a letter or underscore
\item a name starts with "c\if{html}{\out{<digit>}}"
\item a name is a cell reference, such as "C1", "AB6R12"
}
}
}
\examples{
clean_region_name(c("foo-bar", "c19severity", "ad+b 4"))
}
