% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_scopes.R
\name{create_scope_table}
\alias{create_scope_table}
\title{Create a Scope Table}
\usage{
create_scope_table(data, scope, ...)
}
\arguments{
\item{data}{a data frame containing all variable identifiers (join keys) that
belong to the scope.}

\item{scope}{(character) name/ID of the Scope.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> named list containing
additional information about the variable Scope. This will be used as
columns of the Scope Selection Table in the workbook.}
}
\value{
nested tibble decribing the Scope
}
\description{
Create a table that defines variable scopes. This table can be used for
adding a "Scope"-Column (and Selection Rules) to the form item tables and
Scope-Selection-Tables in the workbook.
}
\examples{
scope_data <- scope_data <- data.frame (
  mainform = c("form1", "form1", "form2"),
  varname_col  = c("var1", "var3", "var10")
)
create_scope_table(
  scope_data,
  "my_scope",
  list(Description = "This Scope covers this and that",
       visits = "Baseline")
)
}
