% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\name{format_empty_row}
\alias{format_empty_row}
\title{Format an empty Row in the worksheet}
\usage{
format_empty_row(wb, sheet = current_sheet(), row = 1, doc_width = "G")
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{row}{in which row the As Of Date should be added. Default = 1}

\item{doc_width}{(single character or numeric) Up to which column should the
titles and texts spread? Default = "G"}
}
\value{
A wbWorkbook, invisibly
}
\description{
The row is styled as the style of "text_area", which is the default font
with a white background.
}
\examples{
wb <- openxlsx2::wb_workbook()
style_datadict(wb)
wb$add_worksheet()

format_empty_row(wb, row = 3)
}
