% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themers.R
\name{get_color_theme}
\alias{get_color_theme}
\title{Get Data Dictionary Color Theme}
\usage{
get_color_theme(name = NULL)
}
\arguments{
\item{name}{of color theme. Use \code{\link[=ls_color_themes]{ls_color_themes()}} to list all currently
available themes. If name = NULL (Default), the function returns the
currently active color theme. Use \code{\link[=active_color_theme]{active_color_theme()}} to get the name
of the active color theme.}
}
\value{
named list with the colors defined in the color theme
}
\description{
Get Data Dictionary Color Theme
}
\examples{
get_color_theme()
get_color_theme("purpur")
}
