% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{get_datadict_language}
\alias{get_datadict_language}
\title{Get Language Setting for Data Dictionary}
\usage{
get_datadict_language()
}
\value{
character. A language code that represents the current language
setting that is used for the data  dictionary.
}
\description{
Get Language Setting for Data Dictionary
}
\examples{
get_datadict_language()
}
