% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{get_datadict_language_codes}
\alias{get_datadict_language_codes}
\title{Get available Language Setting for Data Dictionary}
\usage{
get_datadict_language_codes()
}
\value{
character vector with all language codes for the available language
settings of the data dictionary.
}
\description{
Get available Language Setting for Data Dictionary
}
\examples{
get_datadict_language_codes()
}
