% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{get_export_date}
\alias{get_export_date}
\title{Get The Creation Date Time Of The Data Export Directory}
\usage{
get_export_date(data_dir, is_zip = FALSE)
}
\arguments{
\item{data_dir}{Name of the Directory or Zip File in which the SecuTrial or
tsExport data is stored.}

\item{is_zip}{(logical) is files directory a zip file?}
}
\value{
(POSIXt)
}
\description{
This function tries to extract the data creation date time from the
ExportOptions file that is exported with the data.
However if due to some processing steps or other reasons there is no
ExportOptions file found in the data directory, then return the
modification time of the data directory
}
\keyword{internal}
