% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadict_tables.R
\docType{data}
\name{intermediates}
\alias{intermediates}
\title{stdatadict Environment For Intermediate Calculation Results}
\format{
An object of class \code{environment} of length 0.
}
\usage{
intermediates
}
\description{
List the objects currently stored inside this environment:
\code{s(stdatadict::intermediates, all.names = TRUE)}
Access object from this environment:
\verb{stdatadict::intermediates$<ObjName>}
}
\keyword{datasets}
