% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_scopes.R
\name{join_scopes}
\alias{join_scopes}
\title{Add Variable Scopes To Form Item Tables}
\usage{
join_scopes(datadict_tables, scope)
}
\arguments{
\item{datadict_tables}{table list, generated with \code{\link[=create_datadict_tables]{create_datadict_tables()}}}

\item{scope}{"Scope Table" (see Details) or list of Scope tables.}
}
\value{
datadict_tables list where a Scope variable is added to form_item
tables.
}
\description{
This functions joins a "Scope" variable with the form item tables in
"data_dict_tables". If form item tables already have a Scope variable, the
content of already existing and new Scope variable be concatenated.
}
\details{
A Scope table is a nested table with the following format:
\describe{
\item{Scope}{ID String of the variable Scope}
\item{additional Scope Information}{Variables that describe the Scope, such
as Scope Label/Name, Description, etc.}
\item{data}{a table containing variable names and other identifying
variables to which the Scope belongs (joining key variables)}
}
}
\examples{
\dontrun{
datadict_tables <- join_scopes(datadict_tables, datapkg_pop_scope)
}
}
