% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_form_sheets.R
\name{link_form_sheets}
\alias{link_form_sheets}
\title{Link Form Names in the Form Overview}
\usage{
link_form_sheets(wb)
}
\arguments{
\item{wb}{a data dictionary workbook}
}
\value{
workbook with links to form item tables (invisibly)
}
\description{
Link the form names in the "Table"/"Embedding Form" columns in the
Form Overview sheet to their respective form sheets.

Additionally add a link back to the Form Overview to each A1 cell of the
form item sheets.

Note that the form overview and the form sheets must be added to the workook
before calling this function.
}
\examples{
\dontrun{
wb <- wb_workbook()
add_form_overview(wb, datadict_tables = datadict_tables)
add_form_sheets(wb, datadict_tables = datadict_tables)
link_form_sheets(wb)
}
}
