% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themers.R
\name{ls_color_themes}
\alias{ls_color_themes}
\title{List All Available Color Theme Names for Data Dictionary Creation.}
\usage{
ls_color_themes()
}
\value{
vector of available color theme names, invisible
}
\description{
List All Available Color Theme Names for Data Dictionary Creation.
}
\examples{
ls_color_themes()
}
