% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{wb_workbook}
\alias{wb_add_worksheet}
\alias{wb_add_data}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{openxlsx2}{\code{\link[openxlsx2]{wb_add_data}}, \code{\link[openxlsx2]{wb_add_worksheet}}, \code{\link[openxlsx2]{wb_workbook}}}
}}

