% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{set_datadict_language}
\alias{set_datadict_language}
\title{Set Language of the Data Dictionary}
\usage{
set_datadict_language(lang = c("en", "de", stdatadictEnv$available_languages))
}
\arguments{
\item{lang}{a character string that specifies the language code for the
language of the data dictionary. Get a list of available language codes
with \code{\link[=get_datadict_language_codes]{get_datadict_language_codes()}}.}
}
\value{
old language option invisibly.
}
\description{
Set Language of the Data Dictionary
}
\examples{
## Set the language of the data dictionary to german
set_datadict_language("de")
}
