% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themers.R
\name{show_color_theme}
\alias{show_color_theme}
\title{Show Color Theme}
\usage{
show_color_theme(name)
}
\arguments{
\item{name}{of color theme. Use \code{\link[=ls_color_themes]{ls_color_themes()}} to list all currently
available themes.}
}
\value{
name parameter of the function call invisibly
}
\description{
Show Color Theme
}
\examples{
show_color_theme("purpur")
}
