% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdatadict-package.R
\docType{package}
\name{stdatadict-package}
\alias{stdatadict}
\alias{stdatadict-package}
\title{stdatadict: Create Data Dictionary or Variable Selection File from SecuTrial Export}
\description{
This package provides functions that help to create a Data Dictionary or Variable Selection Excel file from the meta data files of an Secutrial csv export and additional files provided by the user.
}
\author{
\strong{Maintainer}: Miriam Rainers \email{miriam.rainers@med.uni-goettingen.de} (\href{https://orcid.org/0009-0003-8385-8968}{ORCID})

}
\keyword{internal}
