% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_datadict.R
\name{style_datadict}
\alias{style_datadict}
\title{Create And Register Data Dictionary Workbook Styles}
\usage{
style_datadict(wb, theme_name = stdatadictEnv$use_color_theme)
}
\arguments{
\item{wb}{a workbook}

\item{theme_name}{Character Name of the color theme that should be used for the data dictionary}
}
\value{
workbook with added styles, invisibly
}
\description{
Create And Register Data Dictionary Workbook Styles
}
\examples{
wb <- openxlsx2::wb_workbook()
style_datadict(wb)
}
