% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themers.R
\name{tint_color}
\alias{tint_color}
\title{Tint or shade a color}
\usage{
tint_color(color, tint)
}
\arguments{
\item{color}{String, either a color name known by R, or the hex rgb code
of the background color.}

\item{tint}{numeric value between -1 and +1. Negative values darkens and
positive values brightens the original color.}
}
\value{
a hex rgb code as string.
}
\description{
Calculate a brighter or darker version of the original color.
}
\examples{
tint_color("purple", 0.4)
tint_color("#FFD530", -0.7)
}
