% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themers.R
\name{use_color_theme}
\alias{use_color_theme}
\title{Set Color Theme That's Gonna Be Used For The Data Dictionary}
\usage{
use_color_theme(name)
}
\arguments{
\item{name}{of color theme. Use \code{\link[=ls_color_themes]{ls_color_themes()}} to list all currently
available themes.}
}
\value{
invisibly the name parameter of this function call
}
\description{
Set Color Theme That's Gonna Be Used For The Data Dictionary
}
\examples{
use_color_theme("purpur")
}
