% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_creators.R
\name{wb_insert_row}
\alias{wb_insert_row}
\title{Insert a row in a sheet (emulated)}
\usage{
wb_insert_row(
  wb,
  sheet = current_sheet(),
  insert_row = 1,
  nrows = 1,
  below = FALSE
)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{sheet}{The worksheet where to insert row(s). Can be the worksheet
index or name. Default: current_sheet()}

\item{insert_row}{numeric, where to insert a row. Default = 1}

\item{nrows}{how many rows should be inserted. Default = 1}

\item{below}{Boolean indicating if the rows inserted below \code{insert_row}.
Default = FALSE}
}
\value{
A wbWorkbook, invisibly
}
\description{
Emulate the function of inserting one or multiple rows in a data sheet.
This is done, by copying and pasting the rows that comes after the inserted
rows and then clearing the insertion rows from data.

Be careful with this function. It will likely mess up the formulas in that
sheet, or that refer to that sheet.
}
\examples{
library(openxlsx2)

# create workbook
wb <- wb_workbook()$
add_worksheet()$
 add_data(x = LifeCycleSavings, row_names = TRUE)

# insert two lines above row 4
wb_insert_row(wb, insert_row = 4, nrows = 2)
}
