% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themers.R
\name{add_color_theme}
\alias{add_color_theme}
\title{Create a New Data Dictionary Color Theme}
\usage{
add_color_theme(
  name,
  template = "purpur",
  use_theme = TRUE,
  bg_color_primary = NULL,
  font_color_primary = NULL,
  bg_color_subtitle = NULL,
  font_color_subtitle = NULL,
  bg_color_h1 = NULL,
  font_color_h1 = NULL,
  bg_color_tablehead = NULL,
  font_color_tablehead = NULL,
  bg_color_visit = NULL,
  font_color_visit = NULL,
  bg_color_tablecol_overview = NULL,
  font_color_tablecol_overview = NULL,
  bg_color_tablecol_items = NULL,
  font_color_tablecol_items = NULL,
  border_color_tablecol_items = NULL,
  font_color_hidden = NULL
)
}
\arguments{
\item{name}{of the new color theme}

\item{template}{name of the color theme from which the new theme inherits
color values. The default is "purpur".}

\item{use_theme}{logical. Should the newly defined theme used for the workbook?
Default = TRUE}

\item{bg_color_primary}{character color name or RGB hex code:
background for document and form titles}

\item{font_color_primary}{character color name or RGB hex code:
font color for document and form titles}

\item{bg_color_subtitle}{character color name or RGB hex code:
background subtitle}

\item{font_color_subtitle}{character color name or RGB hex code:
font color subtitle}

\item{bg_color_h1}{character color name or RGB hex code:
background heading}

\item{font_color_h1}{character color name or RGB hex code:
font color heading}

\item{bg_color_tablehead}{character color name or RGB hex code:
background column names}

\item{font_color_tablehead}{character color name or RGB hex code:
font color column names}

\item{bg_color_visit}{character color name or RGB hex code:
background visit names}

\item{font_color_visit}{character color name or RGB hex code:
font color visit names}

\item{bg_color_tablecol_overview}{character color name or RGB hex code:
background table names in form overview}

\item{font_color_tablecol_overview}{character color name or RGB hex code:
font color table names in form overview}

\item{bg_color_tablecol_items}{character color name or RGB hex code:
background table names in form items}

\item{font_color_tablecol_items}{character color name or RGB hex code:
font color table names in form items}

\item{border_color_tablecol_items}{character color name or RGB hex code:
border table names in form items}

\item{font_color_hidden}{character color name or RGB hex code:
font color hidden forms and items}
}
\value{
new color theme as named list, invisibly
}
\description{
Define a new color theme for the data dictionary workbook, based on the
color definition from a template theme. You only have to define those
parameters that deviate from that template.

The newly defined color theme will be added to the list of available themes
and if \code{use_theme = TRUE} use the new theme for creating the data dictionary
excel workbook.

Use \code{\link[=ls_color_themes]{ls_color_themes()}} to list all currently available themes.
}
\examples{
add_color_theme("my_theme",
                template = "purpur",
                bg_color_primary = "steelblue4",
                font_color_subtitle = "steelblue4",
                bg_color_h1 = "steelblue2",
                font_color_h1 = "#111188")
}
\seealso{
\code{\link[=use_color_theme]{use_color_theme()}} \code{\link[=ls_color_themes]{ls_color_themes()}}
}
