% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themers.R
\name{get_color_theme}
\alias{get_color_theme}
\title{Get Data Dictionary Color Theme}
\usage{
get_color_theme(name)
}
\arguments{
\item{name}{of color theme. Use \code{\link[=ls_color_themes]{ls_color_themes()}} to list all currently
available themes.}
}
\value{
named list with the colors defined in the color theme
}
\description{
Get Data Dictionary Color Theme
}
\examples{
get_color_theme("purpur")
}
