% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read Meta Data Files From A SecuTrial or stExport Directory}
\usage{
read_metadata(data_dir, delim = ";", decimal_mark = ",", encoding = "UTF-8")
}
\arguments{
\item{data_dir}{Name of the Directory or Zip File in which the SecuTrial or
tsExport data is stored.}

\item{delim}{Single character used to separate fields within a record.}

\item{decimal_mark}{Symbol used to indicate the decimal place.
Decimal mark can only be ⁠,⁠ or ..}

\item{encoding}{Default encoding. This only affects how the file is
read - readr always converts the output to UTF-8.}
}
\value{
List of tibbles with meta data read from the SecuTrial or tsExport
directory or zip file.
}
\description{
Read the following csv files from the export directory or zip file:
\itemize{
\item vp.csv or visitplan*.csv
\item vpfs.csv or visitplanforms*.csv
\item fs.csv or forms*.csv
\item qs.csv or questions*.csv
\item cl.csv or cl*.csv
\item is.csv or items*.csv
}

If one of this files isn't present in the data directory, the function exits
with an error message and no files where read.

The column "hidden" will be added to the files: "fs.csv", "qs.csv", and
"is.csv" if it doesn't already exists, because further functions need this
column to be present.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data_dir <- file.choose()
metadata <- read_metadata(data_dir)
\dontshow{\}) # examplesIf}
}
